+++ CHANGELOG.md
@@ -16,9 +16,10 @@
 - Controls. Focus frame isn't rouned on Safari
 - Controls. Attributes `initvalue,min,max` for controls Date & Calendar doesn't work on Safari (Date.parse(yyyy-MM-dd) doesn't work by default)
 - Controls. Controls are not rendered if parsing initvalue is wrong
-- Controls. (Select, Date, Time). Now popup isn't opened if user clears control and gets focus at the same time
 - Controls. Unexpected autofocus on mask-inputs on Safari
 - Controls. Annoying autoselect on touchscreens. Now $options.selectOnFocus is disabled by default
+- Combobox controls (Select, Date, Time). Now popup isn't opened if user clears control and gets focus at the same time
+- Combobox controls (Select, Date, Time). Click on disabled item throws console.error
 - [SpinElement](src/spinElement.ts). Animation for TwinDualRing doesn't work on Safari 14-
 - [SpinElement](src/spinElement.ts). Wrong render on option `fit` on Safari 14-
 - [PopupElement](src/popup/popupElement.ts). Animation for opacity doesn't work on Safari 14-